requirements.txt
requests
pytest
allure-pytest
jsonschema
pydantic
pylint
